refactor: Move tests
