+++ tests/bot/models/commands/test_mentions_list.py
@@ -3,8 +3,7 @@ from uuid import UUID, uuid4
 
 import pytest
 
-from botx import Mention, MentionTypes
-from botx.models.message.mentions import MentionList
+from botx import Mention, MentionList, MentionTypes
 
 
 @pytest.fixture
